+++ src/pyramid_bimt/static/bimt.js
@@ -1,4 +1,4 @@
-/*global document, jQuery */
+/*global document, jQuery, bootbox */
 /*jslint indent: 4, maxlen: 80 */
 
 (function ($) {
@@ -16,12 +16,34 @@
             $(this).text($(this).data('secret'));
         });
 
+        // Enable 'confirmation button' form functionality: a button that
+        // pops-up a confirmation dialog before submitting the form
+        $("form .btn-confirmation").click(function (e) {
+            var $button = $(this),
+                $form = $(this).closest("form");
+
+            e.preventDefault();
+
+            bootbox.confirm($button.attr('value'), function (result) {
+                if (result) {
+                    // the jquery submit() does not include the button in the
+                    // POST so we have to manually append it
+                    $('<input />').attr('type', 'hidden')
+                        .attr('name', $button.attr('name'))
+                        .attr('value', $button.attr('value'))
+                        .appendTo($form);
+
+                    $form.submit();
+                }
+            });
+        });
+
         if ($('.datatable').length > 0 && $('.datatable').dataTable) {
             // Read columns configuration from DOM
             var $table = $('.datatable'),
                 sort_direction = $table.data('sortDescending') ? 'desc' : 'asc',
                 aoColumns = [];
-            $(".table thead th").each(function() {
+            $(".table thead th").each(function () {
                 var $this = $(this);
                 if ($this.data('sortDisabled') === true) {
                     aoColumns.push({ "bSortable": false });