+++ src/pyramid_bimt/tests/test_sanity_check.py
@@ -231,8 +231,8 @@ class TestSanityCheckEmail(unittest.TestCase):
 
         mailer = get_mailer(self.config)
         self.assertEqual(len(mailer.outbox), 1)
-        self.assertEqual(mailer.outbox[0].sender, 'info@bar.com')
-        self.assertEqual(mailer.outbox[0].recipients, ['info@bar.com', ])
+        self.assertEqual(
+            mailer.outbox[0].recipients, ['maintenance@niteoweb.com', ])
         self.assertEqual(
             mailer.outbox[0].subject, u'BIMT sanity check OK')
         self.assertIn(
@@ -247,8 +247,8 @@ class TestSanityCheckEmail(unittest.TestCase):
 
         mailer = get_mailer(self.config)
         self.assertEqual(len(mailer.outbox), 1)
-        self.assertEqual(mailer.outbox[0].sender, 'info@bar.com')
-        self.assertEqual(mailer.outbox[0].recipients, ['info@bar.com', ])
+        self.assertEqual(
+            mailer.outbox[0].recipients, ['maintenance@niteoweb.com', ])
         self.assertEqual(
             mailer.outbox[0].subject, u'BIMT sanity check found 2 warnings')